add new parent exception
