+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/exceptions/InvalidEnvelopeException.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.reform.bulkscanprocessor.exceptions;
 /**
  * Base class for all exceptions that represent invalid envelope situation.
  */
-public abstract class InvalidEnvelopeException extends RuntimeException {
+public abstract class InvalidEnvelopeException extends ProcessorRunTimeException {
 
     public InvalidEnvelopeException(String message) {
         super(message);
