+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/exceptions/InvalidMetafileException.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.reform.bulkscanprocessor.exceptions;
 /**
  * Base class for all exceptions that represent invalid metadata situation.
  */
-public abstract class InvalidMetafileException extends RuntimeException {
+public abstract class InvalidMetafileException extends ProcessorRunTimeException {
 
     public InvalidMetafileException(String message) {
         super(message);
@@ -12,4 +12,9 @@ public abstract class InvalidMetafileException extends RuntimeException {
     public InvalidMetafileException(String message, Throwable cause) {
         super(message, cause);
     }
+
+    public InvalidMetafileException(String message, String detailMessage) {
+        super(message, detailMessage);
+    }
+
 }
