+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/exceptions/PaymentsDisabledException.java
@@ -1,6 +1,6 @@
 package uk.gov.hmcts.reform.bulkscanprocessor.exceptions;
 
-public class PaymentsDisabledException extends RuntimeException {
+public class PaymentsDisabledException extends ProcessorRunTimeException {
     public PaymentsDisabledException(String message) {
         super(message);
     }
