+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/exceptions/ServiceDisabledException.java
@@ -1,6 +1,6 @@
 package uk.gov.hmcts.reform.bulkscanprocessor.exceptions;
 
-public class ServiceDisabledException extends RuntimeException {
+public class ServiceDisabledException extends ProcessorRunTimeException {
     public ServiceDisabledException(String message) {
         super(message);
     }
